commands: move squash code to squash.rs
